clock: Move `types` module to `v2`
