+++ hal/src/thumbv7em/clock.rs
@@ -2,5 +2,3 @@ pub mod v1;
 pub use v1::*;
 
 pub mod v2;
-
-pub mod types;
