+++ hal/src/thumbv7em/clock/v2/dfll.rs
@@ -15,8 +15,10 @@
 //! - in a return value of [`crate::clock::v2::retrieve_clocks`]
 use typenum::{U0, U1};
 
-use crate::clock::types::{Counter, Enabled, PrivateIncrement};
-use crate::clock::v2::{Source, SourceMarker};
+use crate::clock::v2::{
+    types::{Counter, Enabled, PrivateIncrement},
+    Source, SourceMarker,
+};
 use crate::time::{Hertz, U32Ext};
 use crate::typelevel::Sealed;
