+++ hal/src/thumbv7em/clock/v2/gclk.rs
@@ -45,8 +45,10 @@ use crate::pac::NVMCTRL;
 pub use crate::pac::gclk::genctrl::{DIVSEL_A, SRC_A as GclkSourceEnum};
 
 use super::gclkio::NotGclkInput;
-use crate::clock::types::{Counter, Decrement, Enabled, Increment, PrivateIncrement};
-use crate::clock::v2::{Source, SourceMarker};
+use crate::clock::v2::{
+    types::{Counter, Decrement, Enabled, Increment, PrivateIncrement},
+    Source, SourceMarker,
+};
 use crate::pac::gclk::{RegisterBlock, GENCTRL};
 use crate::time::Hertz;
 use crate::typelevel::Sealed;
