+++ hal/src/thumbv7em/clock/v2/gclkio.rs
@@ -62,8 +62,10 @@ use core::marker::PhantomData;
 use seq_macro::seq;
 use typenum::U0;
 
-use crate::clock::types::{Counter, Decrement, Enabled, Increment};
-use crate::clock::v2::{Source, SourceMarker};
+use crate::clock::v2::{
+    types::{Counter, Decrement, Enabled, Increment},
+    Source, SourceMarker,
+};
 use crate::gpio::v2::{self as gpio, AlternateM, AnyPin, Pin, PinId};
 use crate::time::Hertz;
 use crate::typelevel::*;
