+++ hal/src/thumbv7em/clock/v2/types.rs
