+++ hal/src/thumbv7em/clock/v2/xosc32k.rs
@@ -20,8 +20,10 @@ use crate::pac::osc32kctrl::xosc32k::{CGM_A, STARTUP_A};
 
 use crate::pac::osc32kctrl::{RegisterBlock, STATUS, XOSC32K};
 
-use crate::clock::types::{Counter, Enabled};
-use crate::clock::v2::{Source, SourceMarker};
+use crate::clock::v2::{
+    types::{Counter, Enabled},
+    Source, SourceMarker,
+};
 use crate::gpio::v2::{AnyPin, FloatingDisabled, Pin, PA00, PA01};
 use crate::time::{Hertz, U32Ext};
